seo-platform/backend/requirements.txt
# ── Core ──
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9

# ── Database ──
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ── Task Queue ──
celery[redis]==5.4.0
redis==5.0.7

# ── LLM Gateway (Inkwell) ──
litellm==1.40.0
anthropic==0.30.0
openai==1.35.0

# ── Web Research (Inkwell) ──
google-search-results==2.4.2
playwright==1.44.0
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Crawler (KeyGap) ──
scrapy==2.11.2

# ── NLP (KeyGap) ──
spacy==3.7.5
scikit-learn==1.5.1

# ── Plagiarism (Inkwell) ──
datasketch==1.6.4

# ── Utilities ──
httpx==0.27.0
python-dotenv==1.0.1
cryptography==42.0.8
aiofiles==23.2.1
jinja2==3.1.4
python-slugify==8.0.4
textstat==0.7.3

# ── Email ──
aiosmtplib==3.0.1
email-validator==2.1.1

# ── Monitoring ──
prometheus-fastapi-instrumentator==7.0.0

# ── Dev ──
ruff==0.5.0
pytest==8.2.2
pytest-asyncio==0.23.7
